requirements.txt
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.0
billiard==3.6.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
frozenlist==1.3.3
httpie==3.2.1
idna==3.4
iniconfig==2.0.0
kombu==5.2.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
prompt-toolkit==3.0.38
Pygments==2.14.0
PySocks==1.7.1
pytest==7.2.2
pytest-django==4.5.2
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.3
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
